% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download RedCap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the API token stored in the user's .REnviron file.}

\item{redcapUrl}{The URL of the RedCap project.}

\item{redcapReportId}{The ID of the RedCap report to download.}
}
\value{
A tibble containing the contents of the RedCap report.
}
\description{
This function downloads a RedCap report using the provided API token name, URL, and report ID.
}
