% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the parameter p that maximizes the log-likelihood for Bernoulli trials}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of Bernoulli trials (0 or 1).}
}
\value{
The parameter p that maximizes the log-likelihood.
}
\description{
This function takes a vector of Bernoulli trials and calculates the parameter p
that maximizes the log-likelihood.
}
\examples{
data <- c(1, 0, 0, 0, 1, 1, 1)
best_p <- logLikBernoulli(data)
print(best_p)
}
