% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate Data Using Principal Components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame representing the input data.}

\item{npc}{An integer specifying the number of principal components to use
for approximation.}
}
\value{
A numeric matrix of the same dimensions as \code{x}, containing the
approximated data based on the specified number of PCs.
}
\description{
This function approximates the input data \code{x} using a specified number of
principal components (PCs).
}
\examples{
# Generate some sample data
set.seed(123)
x <- matrix(rnorm(100 * 5), nrow = 100, ncol = 5)

# Approximate the data using 3 PCs
x_approx <- pcApprox(x, npc = 3)

}
