% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names}
\usage{
standardizeNames(data, case = "small_camel")
}
\arguments{
\item{data}{A data frame whose variable names are to be standardized.}
}
\value{
A data frame with standardized variable names.
}
\description{
This function standardizes the names of variables in a data frame data to a specified case.
Uses janitor::clean_names and dplyr::rename_with to convert names.
}
\examples{
data <- tibble::tibble('Variable Name' = 1:5, 'Another_Variable' = 6:10)
data_clean <- standardizeNames(data)
}
