% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and Plot Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numeric vector indicating the event status (1 if event occurred, 0 if censored).}

\item{time}{Numeric vector indicating the time to event or time to censoring.}
}
\description{
This function takes two numerical vectors: status (censoring status) and time (time to event or censor),
calculates the survival probability using the Kaplan-Meier estimator, and plots the survival curve.
}
\examples{
# Assuming that 'survival.csv' has been loaded with columns 'time' and 'status'
data <- read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
survCurv(data$status, data$time)
}
