% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Reverse the Centering and Scaling of a Numeric Vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled and centered.}
}
\value{
The original unscaled and uncentered vector.
}
\description{
This function reverses the scaling and centering transformations applied to a numeric vector,
typically done by the \code{scale()} function. It checks for and uses the attributes 'scaled:center' and
'scaled:scale' stored in the scaled vector to reconstruct the original data.
}
\examples{
original <- c(1, 2, 3, 4, 5)
scaled <- scale(original)
unscale(scaled)
# Should return the original vector
}
